% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_equal.R
\name{add_to_equal}
\alias{add_to_equal}
\title{add something to strings to be equal length}
\usage{
add_to_equal(df, add = " ", colname = FALSE, rowname = FALSE)
}
\arguments{
\item{df}{string, vectors, dataframe or matrix}

\item{add}{appengdix}

\item{colname}{logistic value for dataframe or matrix}

\item{rowname}{logistic value for dataframe or matrix}
}
\value{
equal length results
}
\description{
add something to strings to be equal length
}
\examples{
a=c(123,1,24,5,1.22554)
add_to_equal(a)
}
