% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_col.R
\name{all_col}
\alias{all_col}
\title{all column true is true in a data}
\usage{
all_col(data)
}
\arguments{
\item{data}{dataframe or matrix}
}
\value{
character
}
\description{
all column true is true in a data
}
\examples{
any_col(data)
}
