% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_col.R
\name{any_col}
\alias{any_col}
\title{Any column true is true in a dataframe}
\usage{
any_col(dataframe)
}
\arguments{
\item{dataframe}{}
}
\value{
character
}
\description{
Any column true is true in a dataframe
}
\examples{
any_col(dataframe)
}
