% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_row.R
\name{any_row}
\alias{any_row}
\title{Any row true is true in a dataframe}
\usage{
any_row(dataframe)
}
\arguments{
\item{dataframe}{}
}
\value{
character
}
\description{
Any row true is true in a dataframe
}
\examples{
any_row(dataframe)
}
