% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_forEach.R
\name{auc_forEach}
\alias{auc_forEach}
\title{To get auc for each variable}
\usage{
auc_forEach(data, y)
}
\arguments{
\item{data}{data}

\item{y}{classification}
}
\value{
dataframe
}
\description{
To get auc for each variable
}
\examples{
auc_forEach(data,"y")
}
