% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catdf.R
\name{catdf}
\alias{catdf}
\title{Cat Dataframe and Matrix}
\usage{
catdf(df, n = 5)
}
\arguments{
\item{df}{dataframe or matrix}

\item{n}{now of row, default is 5}
}
\value{
dataframe or matrix
}
\description{
Cat Dataframe and Matrix
}
\examples{
\donttest{
df = data.frame(a=c(1,2,3,450,4567,1),
               bc=c(1,2,4,5,2,5),
               c=c(1,2,4,5,2,5),
               d=c(1,2,4,5,2,5),
               e=c(1,2,4,5,2,5),
               f=c(1,2,4,5,2,5))
catdf(df,4)
catdf(df)
}
}
