% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{common}
\alias{common}
\title{Find Common}
\usage{
common(data)
}
\arguments{
\item{data}{must be list}
}
\value{
vectors
}
\description{
Find Common
}
\examples{
\dontrun{
data1=c('a','e','d')
data2=c('a','c','e')
data3=c('a','e','j')
data=list(data1,data2,data3)
common(data)
}
}
