% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_strech.R
\name{flat_strech}
\alias{flat_strech}
\title{To flat and strech data}
\usage{
flat_strech(data, strech = NULL)
}
\arguments{
\item{data}{dataframe}

\item{strech}{string of name}
}
\value{
a dataframe
}
\description{
To flat and strech data
}
\examples{
flat_strech(data)
flat_strech(data,"strech")
}
