% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapesEqualsGrapes.R
\name{\%==\%}
\alias{\%==\%}
\title{return locations which b equals a in b}
\usage{
a \%==\% b
}
\arguments{
\item{a}{object}

\item{b}{character}
}
\value{
vector or list
}
\description{
return locations which b equals a in b
}
\examples{
"M" \%==\% c("M","m","M1","M")
}
