% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapesNotEqualsGrapes.R
\name{\%!=\%}
\alias{\%!=\%}
\title{return locations which b does not equal a in b}
\usage{
a \%!=\% b
}
\arguments{
\item{a}{object}

\item{b}{character}
}
\value{
vector
}
\description{
return locations which b does not equal a in b
}
\examples{
"a" \%!=\% c("a","ab","a","b","c")
c("a","b") \%!=\% c("a","ab","a","b","c")
}
