% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapesSimilarEqualsGrapes.R
\name{\%s=\%}
\alias{\%s=\%}
\title{similarly match by grep}
\usage{
a \%s=\% b
}
\arguments{
\item{a}{a}

\item{b}{b}
}
\value{
vector
}
\description{
similarly match by grep
}
\examples{
1 "\%s=\%" c(1,12,3)
}
