% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_Add_Symbol.R
\name{inner_Add_Symbol}
\alias{inner_Add_Symbol}
\title{To inner join character by some symbol}
\usage{
inner_Add_Symbol(character, symbol = "+")
}
\arguments{
\item{character}{character}

\item{symbol}{symbol}
}
\value{
character
}
\description{
To inner join character by some symbol
}
\examples{
inner_Add_Symbol(c(1,2,4),"+")
inner_Add_Symbol(c(1),"+")
}
