% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Left.R
\name{Left}
\alias{Left}
\title{cut out a string from left}
\usage{
Left(x, n)
}
\arguments{
\item{x}{string}

\item{n}{length}
}
\value{
string
}
\description{
cut out a string from left
}
\examples{
Left("abcd",3)
}
