% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mid.R
\name{Mid}
\alias{Mid}
\title{cut out a string from start and length}
\usage{
Mid(x, start, n)
}
\arguments{
\item{x}{string}

\item{start}{start}

\item{n}{length}
}
\value{
string
}
\description{
cut out a string from start and length
}
\examples{
Mid("abcd",2,2)
}
