% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nchar.R
\name{Nchar}
\alias{Nchar}
\title{number of characters}
\usage{
Nchar(x)
}
\arguments{
\item{x}{vectors, dataframe or matrix}
}
\value{
number
}
\description{
number of characters
}
\examples{
Nchar(12)
}
