% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normaldistr_test.R
\name{Normaldistr_test}
\alias{Normaldistr_test}
\title{To do normal distrbution test}
\usage{
Normaldistr_test(x)
}
\arguments{
\item{x}{must be numeric or integer}
}
\value{
test results
}
\description{
To do normal distrbution test
}
\examples{
Normaldistr_test(rnorm(100))
}
