% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{To plot roc curve}
\usage{
plot_roc(score, y, YIndex = TRUE)
}
\arguments{
\item{score}{numeric score}

\item{y}{class}

\item{YIndex}{yue deng point}
}
\value{
plot
}
\description{
To plot roc curve
}
\examples{
plot_roc(score,y)
}
