% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{Power}
\alias{Power}
\title{To caculate the power of some data}
\usage{
Power(data, power)
}
\arguments{
\item{data}{data}

\item{power}{power}
}
\value{
data
}
\description{
To caculate the power of some data
}
\examples{
Power(c(1,2,4),2)
}
