% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel.R
\name{read_excel}
\alias{read_excel}
\title{To read excel by xlsx::read.xlsx}
\usage{
read_excel(filename, sheetIndex = 1)
}
\arguments{
\item{filename}{excel name}

\item{sheetIndex}{1 as default}
}
\value{
an excel
}
\description{
To read excel by xlsx::read.xlsx
}
\examples{
read_excel("1.xlsx")
}
