% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Text.R
\name{read_Text}
\alias{read_Text}
\title{To read text in rstudio source to dataframe}
\usage{
read_Text(text)
}
\arguments{
\item{text}{}
}
\value{
dataframe
}
\description{
To read text in rstudio source to dataframe
}
\examples{
read_text()
}
