% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_coxph.R
\name{reg_coxph}
\alias{reg_coxph}
\title{regress for coxph from survival package}
\usage{
reg_coxph(data, time, ending, x)
}
\arguments{
\item{data}{data}

\item{time}{time}

\item{ending}{ending or status}

\item{x}{x, missing means all the rest variables}
}
\value{
list
}
\description{
regress for coxph from survival package
}
\examples{
reg_coxph(data,"time","ending",c("x1","x2"))
}
