% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_linear.R
\name{reg_linear}
\alias{reg_linear}
\title{regress for linear regressiong using lm}
\usage{
reg_linear(data, y, x)
}
\arguments{
\item{data}{data}

\item{y}{y}

\item{x}{x, missing means all the rest variables}
}
\value{
list
}
\description{
regress for linear regressiong using lm
}
\examples{
reg_linear(data,"y","x")
}
