% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_logit.R
\name{reg_logistic}
\alias{reg_logistic}
\title{regress for logistic regression using glm}
\usage{
reg_logistic(data, y, x)
}
\arguments{
\item{data}{data}

\item{y}{y}

\item{x}{x, missing means all the rest variables}
}
\value{
list
}
\description{
regress for logistic regression using glm
}
\examples{
reg_logistic(data,"y","x")
}
