% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.empty.unite.in.list.R
\name{remove.empty.unite.in.list}
\alias{remove.empty.unite.in.list}
\title{To delet empty unit in list}
\usage{
remove.empty.unite.in.list(TRFdata)
}
\arguments{
\item{TRFdata}{list with only 1 degree}
}
\value{
list without empty
}
\description{
To delet empty unit in list
}
\examples{
remove.empty.unite.in.list(list)
}
