% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_objects.R
\name{remove_objects}
\alias{remove_objects}
\title{remove all objects in Global Environment}
\usage{
remove_objects()
}
\value{
nothind
}
\description{
remove all objects in Global Environment
}
\examples{
remove_objects()
}
