% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace}
\alias{replace}
\title{replace}
\usage{
replace(data, from, to, pattern)
}
\arguments{
\item{data}{data}

\item{from}{frmo}

\item{to}{to}
}
\value{
data replaced
}
\description{
using gsub to be more power
}
\examples{
replace(222,2,1)
}
