% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace0.R
\name{replace0}
\alias{replace0}
\title{replace}
\usage{
replace0(data, from)
}
\arguments{
\item{data}{data}

\item{from}{frmo}

\item{to}{to}
}
\value{
data replaced
}
\description{
using gsub to be more power
}
\examples{
replace(222,2)
}
