% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_toLong.R
\name{reshape_toLong}
\alias{reshape_toLong}
\title{reshape wide to long}
\usage{
reshape_toLong(data, id, j, value.var.prefix)
}
\arguments{
\item{data}{data}

\item{id}{id}

\item{j}{j new colname}

\item{value.var.prefix}{value vars}
}
\value{
data.frame
}
\description{
reshape wide to long
}
\examples{
reshape_toLong(data,id,j,value.var.prefix)
}
