% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_toWide.R
\name{reshape_toWide}
\alias{reshape_toWide}
\title{To reshape long to wide}
\usage{
reshape_toWide(data, id, j, value.var.prefix)
}
\arguments{
\item{data}{data}

\item{id}{id}

\item{j}{j}

\item{value.var.prefix}{value variable prefix}
}
\value{
dataframe
}
\description{
To reshape long to wide
}
\examples{
reshape_toWide(data,id,j,value.var.prefix)
}
