% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reverse.R
\name{Reverse}
\alias{Reverse}
\title{Reverse string to the opposite order}
\usage{
Reverse(data)
}
\arguments{
\item{data}{data}
}
\value{
string
}
\description{
Reverse string to the opposite order
}
\examples{
Reverse("abc123")
Reverse(c("abc123","if"))
}
