% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Right.R
\name{Right}
\alias{Right}
\title{cut out a string from right}
\usage{
Right(x, n)
}
\arguments{
\item{x}{string}

\item{n}{length}
}
\value{
string
}
\description{
cut out a string from right
}
\examples{
Right("abcd",3)
}
