% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskFactor_plot.linear.R
\name{RiskFactor_plot.linear}
\alias{RiskFactor_plot.linear}
\title{To plot Risk Factor plot for linear regression}
\usage{
RiskFactor_plot.linear(data, y, patient.names = "name",
  scatter.margin = c(low = 0, left = 4, top = 3, right = 8),
  scatter.color = c(Prediction = "red", Reality = "black"),
  scatter.location = c(legend.x = 10, legend.y = 20),
  scatter.size = c(axis = 1.5, lab = 1.5, points = 1.5, legend = 1.6,
  cutoff = 1.9), heatmap.cellheight = 20, heatmap.cellwidth = 10,
  cluster_cols = FALSE, cluster_rows = FALSE,
  heatmap.color = c("green", "black", "red"), heatmap.size = c(row =
  15, column = 15, legend = 10))
}
\arguments{
\item{data}{the data you want to plot}

\item{y}{the purpose var}

\item{patient.names}{patient id}

\item{scatter.margin}{scatter.margin}

\item{scatter.color}{scatter.color}

\item{scatter.location}{scatter.location}

\item{scatter.size}{scatter.size}

\item{heatmap.cellheight}{heatmap.cellheight}

\item{heatmap.cellwidth}{heatmap.cellwidth}

\item{cluster_cols}{cluster_cols}

\item{cluster_rows}{cluster_rows}

\item{heatmap.color}{heatmap.color}

\item{heatmap.size}{heatmap.size}
}
\value{
four pictures
}
\description{
To plot Risk Factor plot for linear regression
}
\examples{
RiskFactor_plot.linear(
           data=linearData,
           y="y2",
           patient.names="name",

           scatter.margin = c(low=0,left=4,top=3,right=8),
           scatter.color=c(Prediction="red",Reality="black"),
           scatter.location=c(legend.x=10,legend.y=20),
           scatter.size=c(axis=1.5,lab=1.5,
                          points=1.5,legend=1.6,
                          cutoff=1.9),

           heatmap.cellheight=20,
           heatmap.cellwidth=10,
           cluster_cols = FALSE,
           cluster_rows = FALSE,
           heatmap.color=c("green", "black", "red"),
           heatmap.size=c(row=15,column=15,legend=10)
         )
}
