% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskFactor_plot.logistic.R
\name{RiskFactor_plot.logistic}
\alias{RiskFactor_plot.logistic}
\title{To plot Risk Factor plot for logistic regression}
\usage{
RiskFactor_plot.logistic(data, y, patient.names = NULL,
  scatter.margin = c(low = 0, left = 4, top = 3, right = 8),
  scatter.xlab = "patient", scatter.ylab = "Risk Score",
  scatter.color = c(code1 = "red", code0 = "black"),
  scatter.legend.text = c(code1 = "Death", code0 = "Alive"),
  scatter.location = c(cutoff.x = 4, cutoff.y = 0.5, , legend.x = 10,
  legend.y = 0.8), scatter.size = c(axis = 1.5, lab = 1.5, , points =
  1.5, legend = 1.6, , cutoff = 1.9), heatmap.cellheight = 20,
  heatmap.cellwidth = 10, cluster_cols = FALSE, cluster_rows = FALSE,
  heatmap.color = c("green", "black", "red"), heatmap.size = c(row =
  15, column = 15, legend = 10))
}
\arguments{
\item{data}{the data you want to plot}

\item{y}{the purpose var}

\item{patient.names}{patient id}

\item{scatter.margin}{scatter.margin}

\item{scatter.xlab}{xlab}

\item{scatter.ylab}{ylab}

\item{scatter.color}{scatter.color}

\item{scatter.legend.text}{text}

\item{scatter.location}{scatter.location}

\item{scatter.size}{scatter.size}

\item{heatmap.cellheight}{heatmap.cellheight}

\item{heatmap.cellwidth}{heatmap.cellwidth}

\item{cluster_cols}{cluster_cols}

\item{cluster_rows}{cluster_rows}

\item{heatmap.color}{heatmap.color}

\item{heatmap.size}{heatmap.size}
}
\value{
four pictures
}
\description{
To plot Risk Factor plot for logistic regression
}
\examples{
RiskFactor_plot.logistic(
                         data=logisticData,
                         y="y2",
                         patient.names="name",

                         scatter.margin = c(low=0,left=4,top=3,right=8),
                         scatter.xlab="patient",
                         scatter.ylab ="Risk Score",
                         scatter.color=c(code1="red",code0="black"),
                         scatter.legend.text=c(code1="Death",code0="Alive"),
                         scatter.location=c(cutoff.x=4,cutoff.y=0.5,
                                            legend.x=10,legend.y=0.8),

                         scatter.size=c(axis=1.5,lab=1.5,
                                        points=1.5,legend=1.6,
                                        cutoff=1.9),

                         heatmap.cellheight=20,
                         heatmap.cellwidth=10,
                         cluster_cols = FALSE,
                         cluster_rows = FALSE,
                         heatmap.color=c("green", "black", "red"),
                         heatmap.size=c(row=15,column=15,legend=10)
                        )
}
