% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskFactor_plot.survival.R
\name{RiskFactor_plot.survival}
\alias{RiskFactor_plot.survival}
\title{To plot Risk factor plot for survival plot}
\usage{
RiskFactor_plot.survival(data, time, status, patient.names = NULL,
  scatter.legend.text = c(code1 = "Death", code0 = "Alive"),
  scatter.ylab = c(ylab1 = "Risk Score", ylab2 = "Survival Time"),
  scatter.margin = c(low = 1, left = 7.5, top = 4, right = 19),
  scatter.color = c(code0 = "#00dae0", code1 = "#ff9289"),
  scatter.location = c(cutoff.x = 23, cutoff.y = 95, legend1.x = 35,
  legend1.y = 100, legend2.x = 35, legend2.y = 40), scatter.size = c(axis
  = 1.5, lab = 1.5, points = 1.5, legend = 1.6, cutoff = 1.9),
  heatmap.cellheight = 20, heatmap.cellwidth = 15,
  cluster_cols = FALSE, cluster_rows = FALSE,
  heatmap.color = c("green", "black", "red"), heatmap.size = c(row =
  15, column = 15, legend = 10))
}
\arguments{
\item{data}{dataframe}

\item{time}{time}

\item{status}{status}

\item{patient.names}{patient.names}

\item{scatter.legend.text}{scatter.legend.text}

\item{scatter.ylab}{scatter.ylab}

\item{scatter.margin}{scatter.margin}

\item{scatter.color}{scatter.color}

\item{scatter.location}{scatter.location}

\item{scatter.size}{scatter.size}

\item{heatmap.cellheight}{heatmap.cellheight}

\item{heatmap.cellwidth}{heatmap.cellwidth}

\item{cluster_cols}{cluster_cols}

\item{cluster_rows}{cluster_rows}

\item{heatmap.color}{heatmap.color}

\item{heatmap.size}{heatmap.size}
}
\value{
pictures
}
\description{
To plot Risk factor plot for survival plot
}
\examples{
RiskFactor_plot.survival(
                       data=survivalData,
                       time="time",
                       status="status",
                       patient.names="name",

                       scatter.legend.text=c(code1="Death",code0="Alive"),
                       scatter.ylab =c(ylab1="Risk Score",ylab2="Survival Time"),
                       scatter.margin = c(low=1,left=7.5,top=4,right=19),
                       scatter.color=c(code0="#00dae0",code1="#ff9289"),
                       scatter.location=c(cutoff.x=23,cutoff.y=95,
                                          legend1.x=35,legend1.y=100,
                                          legend2.x=35,legend2.y=40),
                       scatter.size=c(axis=1.5,lab=1.5,
                                      points=1.5,legend=1.6,
                                      cutoff=1.9),

                       heatmap.cellheight=20,
                       heatmap.cellwidth=15,
                       cluster_cols = FALSE,
                       cluster_rows = FALSE,
                       heatmap.color=c("green", "black", "red"),
                       heatmap.size=c(row=15,column=15,legend=10)
                     )
}
