% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_col.R
\name{scale_col}
\alias{scale_col}
\title{To scale every column seperately}
\usage{
scale_col(dataframe)
}
\arguments{
\item{dataframe}{you want to scale}
}
\value{
a dataframe
}
\description{
To scale every column seperately
}
\examples{
scale_col(dataframe)
}
