% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_row.R
\name{scale_row}
\alias{scale_row}
\title{To scale every row seperately}
\usage{
scale_row(dataframe)
}
\arguments{
\item{dataframe}{you want to scale}
}
\value{
a dataframe
}
\description{
To scale every row seperately
}
\examples{
scale_row(dataframe)
}
