% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_byonecolumn.R
\name{spread_byonecolumn}
\alias{spread_byonecolumn}
\title{To spread by one column}
\usage{
spread_byonecolumn(data, bycolumn, joint)
}
\arguments{
\item{data}{data frame}

\item{bycolumn}{column index, number}

\item{joint}{joint between words in by column data}
}
\value{
data.frame
}
\description{
To spread by one column
}
\examples{
spread_byonecolumn(data,bycolumn,",")
}
