% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_col.R
\name{sum_col}
\alias{sum_col}
\title{sum for each column}
\usage{
sum_col(data)
}
\arguments{
\item{data}{dataframe or matrix}
}
\value{
vector
}
\description{
sum for each column
}
\examples{
sum_col(data)
}
