% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{To get summary table}
\usage{
summary_table(data, group)
}
\arguments{
\item{data}{data}

\item{group}{variable names}
}
\value{
dataframe
}
\description{
To get summary table
}
\examples{
summary_table(data,"sex")
}
