% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_1_stack_n.R
\name{unique_1_stack_n}
\alias{unique_1_stack_n}
\title{To unique and stack dataframe}
\usage{
unique_1_stack_n(data = dataframe, unique = 1, stack = 2,
  joint = ",")
}
\arguments{
\item{data}{data}

\item{unique}{number}

\item{stack}{number}

\item{joint}{joint of different words}
}
\value{
a dataframe
}
\description{
To unique and stack dataframe
}
\examples{
unique_1_stack_n(data,1,2)
unique_1_stack_n(data,1:2,3)
}
