% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivarReg_logistic_adjust.R
\name{UnivarReg_logistic_adjust}
\alias{UnivarReg_logistic_adjust}
\title{To do UnivarReg of logistic for adjust vars}
\usage{
UnivarReg_logistic_adjust(data, y, adjust)
}
\arguments{
\item{data}{data}

\item{y}{y}

\item{adjust}{adjust vars}
}
\value{
dataframe
}
\description{
To do UnivarReg of logistic for adjust vars
}
\examples{
UnivarReg_logistic_adjust(data,"y",c("adj1","adj2"))
}
