% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniverReg_logistcic.R
\name{UniverReg_logistcic}
\alias{UniverReg_logistcic}
\title{Univariable logistci regression}
\usage{
UniverReg_logistcic(shuju, y)
}
\arguments{
\item{shuju}{data}

\item{y}{y}
}
\value{
dataframe
}
\description{
Univariable logistci regression
}
\examples{
UniverReg_logistcic(shuju,y)
}
