% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_excel.R
\name{write_excel}
\alias{write_excel}
\title{To write an excle}
\usage{
write_excel(dataframe, xlsename)
}
\arguments{
\item{dataframe}{dataframe}

\item{xlsename}{excle name}
}
\value{
an excel
}
\description{
To write an excle
}
\examples{
write_excel(dataframe)
dataframe(dataframe,"data.xlsx")
}
